productint to long

+++ product.java
@@ -1,5 +1,5 @@
 class product{
-    int a=10;
+    long a=10;
     public static void main(String [] args){
         System.out.println("the number is: "+a)
     }
